have UtcDateTimeSerializer throw an exception
UtcDateTimeSerializer.FromString currently returns DateTime.MinValue
on invalid dates. This is a clear bug, as DateTime.MinValue is not an
UTC date, it is rather set as unspecified (DateTime.MinValue.Kind == Unspecified).
This causes problems later on when SessionSchedule.IsNewSession
is invoked on those dates.

These changes have UtcDateTimeSerializer.FromString not have a
DateTime.MinValue fallback, and its key use
(FileStore.InitializeSessionCreateTime) throw a wrapped exception
immediately on an invalid date (with the invalid date string and file).
This makes diagnosing invalid creation dates in session files much
easier.